+++ lib/src/lints/avoid_unused_parameters/models/avoid_unused_parameters.dart
@@ -2,18 +2,18 @@ import 'package:solid_lints/src/common/parameters/excluded_identifiers_list_para
 
 /// A data model class that represents the "avoid returning widgets" input
 /// parameters.
-class AvoidReturningWidgetsParameters {
+class AvoidUnusedParameters {
   /// A list of methods that should be excluded from the lint.
   final ExcludedIdentifiersListParameter exclude;
 
-  /// Constructor for [AvoidReturningWidgetsParameters] model
-  AvoidReturningWidgetsParameters({
+  /// Constructor for [AvoidUnusedParameters] model
+  AvoidUnusedParameters({
     required this.exclude,
   });
 
   /// Method for creating from json data
-  factory AvoidReturningWidgetsParameters.fromJson(Map<String, dynamic> json) {
-    return AvoidReturningWidgetsParameters(
+  factory AvoidUnusedParameters.fromJson(Map<String, dynamic> json) {
+    return AvoidUnusedParameters(
       exclude: ExcludedIdentifiersListParameter.defaultFromJson(json),
     );
   }