+++ lib/src/lints/cyclomatic_complexity/models/cyclomatic_complexity_parameters.dart
@@ -1,7 +1,12 @@
+import 'package:solid_lints/src/common/parameters/excluded_identifiers_list_parameter.dart';
+
 /// Cyclomatic complexity metric limits configuration.
 class CyclomaticComplexityParameters {
   /// Threshold cyclomatic complexity level, exceeding it triggers a warning.
   final int maxComplexity;
+
+  /// A list of methods that should be excluded from the lint.
+  final ExcludedIdentifiersListParameter exclude;
 
   /// Reference: NIST 500-235 item 2.5
   static const _defaultMaxComplexity = 10;
@@ -9,11 +14,13 @@ class CyclomaticComplexityParameters {
   /// Constructor for [CyclomaticComplexityParameters] model
   const CyclomaticComplexityParameters({
     required this.maxComplexity,
+    required this.exclude,
   });
 
   /// Method for creating from json data
   factory CyclomaticComplexityParameters.fromJson(Map<String, Object?> json) =>
       CyclomaticComplexityParameters(
         maxComplexity: json['max_complexity'] as int? ?? _defaultMaxComplexity,
+        exclude: ExcludedIdentifiersListParameter.defaultFromJson(json),
       );
 }